csrf/pyproject.toml
[tool.poetry]
name = "csrf"
version = "0.1.0"
description = ""
authors = ["Ben Simms <ben@bensimms.moe>"]

[tool.poetry.dependencies]
python = "^3.9"
starlette = "^0.14.1"
Jinja2 = "^2.11.2"
WTForms = "^2.3.3"
uvicorn = "^0.13.3"
python-dotenv = "^0.15.0"
itsdangerous = "^1.1.0"
python-multipart = "^0.0.5"
aiohttp = "^3.7.3"
furl = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
